+++ qiskit/test/mock/fake_tokyo.py
@@ -16,7 +16,11 @@
 Fake Tokyo device (20 qubit).
 """
 
-from qiskit.providers.models import GateConfig, QasmBackendConfiguration
+import os
+import json
+
+from qiskit.providers.models import (GateConfig, QasmBackendConfiguration,
+                                     BackendProperties)
 from .fake_backend import FakeBackend
 
 
@@ -62,3 +66,12 @@ class FakeTokyo(FakeBackend):
         )
 
         super().__init__(configuration)
+
+    def properties(self):
+        """Returns a snapshot of device properties as recorded on 8/30/19.
+        """
+        dirname = os.path.dirname(__file__)
+        filename = "props_tokyo.json"
+        with open(os.path.join(dirname, filename), "r") as f_prop:
+            props = json.load(f_prop)
+        return BackendProperties.from_dict(props)